requirements.txt
# HTTP client with better performance
curl_cffi==0.11.3

# TON blockchain operations
tonutils==0.4.5

# Logging
loguru==0.7.3

# Async HTTP operations
aiohttp>=3.8.0

# Environment variables
python-dotenv>=1.0.0

# Rate limiting
asyncio-throttle>=1.0.0

# CAPTCHA solving
anticaptchaofficial>=1.0.46

# Telegram Stars payments via Telethon
telethon>=1.34.0

# JWT token refresh via Selenium
selenium>=4.15.0
webdriver-manager>=4.0.0

# Process management
psutil>=5.9.0

# Additional dependencies
aiohttp>=3.9.0